requirements.txt
Flask==1.1.2
Jinja2==2.11.2
Werkzeug==1.0.1
gunicorn==20.0.4
numpy==1.18.1
Pillow==7.1.2
torchvision==0.9.0

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu